refactor[#523] : widgets/map/ui 타입 수정
