+++ src/widgets/map/ui/placeMarkingListBottomSheet.tsx
@@ -2,18 +2,11 @@ import { useLocation, useNavigate } from "react-router-dom";
 import { useQueryClient } from "@tanstack/react-query";
 import { useMap } from "@vis.gl/react-google-maps";
 import { MarkingItem, MarkingItemSkeleton } from "@/widgets/marking/ui";
-import {
-  Bounds,
-  useMapQueryParams,
-  usePlaceQueryParams,
-} from "@/features/map/hooks";
+import { useMapQueryParams, usePlaceQueryParams } from "@/features/map/hooks";
 import { SortTypeFilter } from "@/features/marking/ui";
-import { LatLng } from "@/entities/auth/api";
-import {
-  SortType,
-  useGetMarkingDetail,
-  useGetMarkingList,
-} from "@/entities/marking/api";
+import type { Bounds, LatLng } from "@/entities/map/types/client";
+import { useGetMarkingDetail, useGetMarkingList } from "@/entities/marking/api";
+import type { SortType } from "@/entities/marking/types/server";
 import { EmptyMarkingThumbnailGrid } from "@/entities/marking/ui";
 import { ROUTER_PATH } from "@/shared/constants";
 import { useInfiniteScroll } from "@/shared/lib";
@@ -93,7 +86,7 @@ const PlaceMarkingList = ({
   const queryClient = useQueryClient();
   const map = useMap();
 
-  const handleRegionClick = ({ lat, lng }: LatLng) => {
+  const handleRegionClick = ({ lat, lng }: NonNullableObject<LatLng>) => {
     map.setCenter({
       lat,
       lng,
